Separating testable components from execution
